backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
psycopg[binary,pool]==3.2.1
python-dotenv==1.0.1
pydantic==2.9.2
httpx==0.27.2
langgraph==0.2.34
langchain-core==0.3.10
huggingface-hub==0.25.1
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.149.0
qdrant-client==1.12.0
sentence-transformers==3.2.0
tavily-python==0.5.0
python-multipart==0.0.26
